- refs #81965: Added support for scenarios in the Download section.
