+++ apps/src/context/climate-variable-provider.tsx
@@ -4,7 +4,7 @@ import {
 	useAppDispatch,
 	useAppSelector
 } from "@/app/hooks";
-import ClimateVariableContext, { ClimateVariableContextType } from "@/hooks/use-climate-variable";
+import ClimateVariableContext from "@/hooks/use-climate-variable";
 import { PostData } from "@/types/types";
 import {
 	setClimateVariable,
@@ -17,6 +17,20 @@ import {
 	ClimateVariableConfigInterface,
 	ClimateVariableInterface, InteractiveRegionOption
 } from "@/types/climate-variable-interface";
+
+export type ClimateVariableContextType = {
+	climateVariable: ClimateVariableInterface | null;
+	selectClimateVariable: (variable: PostData) => void;
+	setVersion: (version: string) => void;
+	setScenario: (scenario: string) => void;
+	setAnalyzeScenarios: (analyzeScenarios: string[]) => void;
+	setThreshold: (threshold: string) => void;
+	setInteractiveRegion: (interactiveRegion: InteractiveRegionOption) => void;
+	setFrequency: (frequency: string) => void;
+	setAnalysisFieldValue: (key: string, value: string | null) => void;
+	setAveragingType: (type: AveragingType) => void;
+	setDateRange: (dates: string[]) => void;
+}
 
 type ClassMapType = Record<string, new (arg: ClimateVariableConfigInterface) => ClimateVariableInterface>;
 
@@ -92,6 +106,12 @@ export const ClimateVariableProvider: React.FC<{ children: React.ReactNode }> = 
 		}));
 	}, [dispatch]);
 
+	const setAnalyzeScenarios = useCallback((scenarios: string[]) => {
+		dispatch(updateClimateVariable({
+			analyzeScenarios: scenarios
+		}));
+	}, [dispatch]);
+
 	const setScenario = useCallback((scenario: string) => {
 		dispatch(updateClimateVariable({
 			scenario
@@ -140,6 +160,7 @@ export const ClimateVariableProvider: React.FC<{ children: React.ReactNode }> = 
 		selectClimateVariable,
 		setVersion,
 		setScenario,
+		setAnalyzeScenarios,
 		setThreshold,
 		setInteractiveRegion,
 		setFrequency,
